refactor: move highlights feed into organism component
