+++ pages/feed/[id].tsx
@@ -1,8 +1,6 @@
 import { GetServerSidePropsContext } from "next";
-
-import Feed from "./index";
-
-export default Feed;
+import Feeds from "components/organisms/HighlightsFeed/feed";
+import ProfileLayout from "layouts/profile";
 
 export const getServerSideProps = async (context: HighlightSSRPropsContext) => {
   return await handleHighlightSSR(context);
@@ -33,3 +31,7 @@ export async function handleHighlightSSR({ params }: GetServerSidePropsContext<{
     props: { highlight },
   };
 }
+
+Feeds.PageLayout = ProfileLayout;
+
+export default Feeds;
